// Code/src/app/shared/model/RegionModel.ts
export class RegionModel {
    id: number;
    regionCode: string;
    regionName: string;
    address: string;
    cityId: number;
    pincode: number;
    stateId: number;
    phoneNo: string;
    mobileNo: string;
    emailId: string;
    cashBalanceLimit: string;
    bankBalanceLimit: string;
    defaultCashLedger: number;
    defaultBankLegder: number;
}